# api/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API for my personal dev website",
		"title": "Dev Website API",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"name": "Matheo Galuba",
			"url": "https://dev.matheo-galuba.com/contact",
			"email": "matheo.galu56@gmail.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		},
		"version": "1.0"
	},
	"host": "localhost:8080",
	"basePath": "/api",
	"paths": {
		"/auth/login": {
			"post": {
				"description": "Login or register a user by email",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["auth"],
				"summary": "Login or register a user",
				"parameters": [
					{
						"description": "Email",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.LoginBody"
						}
					}
				],
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			}
		},
...
		"/auth/refresh": {
			"post": {
				"description": "Refresh a user's access token by refresh token",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["auth"],
				"summary": "Refresh a user's access token",
				"parameters": [
					{
						"description": "Refresh token",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.VerifyBody"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				}
			}
		},
		"/auth/verify": {
			"post": {
				"description": "Verify a user's email by token",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["auth"],
				"summary": "Verify a user's email",
				"parameters": [
					{
						"description": "Token",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.VerifyBody"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/models.User"
						}
					},
					"400": {
						"description": "Bad Request"
					}
				}
			}
		},
		"/auth/{provider}": {
			"post": {
				"description": "Exchange an authorization code for the identity of its owner and attach it to the user having the same email, creating the user if needed. When called with a bearer token, the identity must have the email of the authenticated user.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["auth"],
				"summary": "Login, register or link an identity with an OAuth provider",
				"parameters": [
					{
						"enum": ["github", "google"],
						"type": "string",
						"description": "Identity provider",
						"name": "provider",
						"in": "path",
						"required": true
					},
					{
						"description": "Authorization code and the redirect URI it was issued for",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.OAuthLoginBody"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"409": {
						"description": "Conflict",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"controllers.LoginBody": {
			"type": "object",
			"required": ["email"],
			"properties": {
				"email": {
					"type": "string"
				}
			}
		},
		"controllers.OAuthLoginBody": {
			"type": "object",
			"required": ["code"],
			"properties": {
				"code": {
					"type": "string"
				},
				"redirectUri": {
					"type": "string"
				}
			}
		},
...
		"controllers.VerifyBody": {
			"type": "object",
			"required": ["token"],
			"properties": {
				"token": {
					"type": "string"
				}
			}
		},
...
		"models.Identity": {
			"type": "object",
			"properties": {
				"avatarUrl": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"lastLogin": {
					"type": "string"
				},
				"linkedAt": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"profileUrl": {
					"type": "string"
				},
				"provider": {
					"type": "string",
					"enum": ["email", "github", "google"]
				},
				"providerId": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			}
		},
...
		"models.Passkey": {
			"type": "object",
			"properties": {
				"createdAt": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"lastUsed": {
					"type": "string"
				},
				"name": {
					"type": "string"
				}
			}
		},
...
		"models.User": {
			"type": "object",
			"properties": {
				"createdAt": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"flavour": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"identities": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.Identity"
					}
				},
				"lastLogin": {
					"type": "string"
				},
				"lastRefresh": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"passkeys": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.Passkey"
					}
				},
				"profilePicture": {
					"type": "string"
				}
			}
		}
	},
	"securityDefinitions": {
		"Bearer": {
			"description": "Bearer token",
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}